% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.scale.R
\name{center.scale}
\alias{center.scale}
\title{Center and scale}
\usage{
center.scale(X)
}
\arguments{
\item{X}{a numeric matrix (or data frame) to center and scaled.
\code{NA}s not allowed.}
}
\value{
\code{center.scale} returns a centered and scaled matrix.
}
\description{
Center and scale a dataset.
}
\examples{
data("nutrimouse")
\dontrun{
 nutrimouse.sc <- center.scale(nutrimouse$gene)
}

}
\seealso{
\code{\link{compute.kernel}}, \code{\link{combine.kernels}}
}
\author{
Celine Brouard <celine.brouard@inrae.fr>
Jerome Mariette <jerome.mariette@inrae.fr>
Nathalie Vialaneix <nathalie.vialaneix@inrae.fr>
}
