% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.pca.permute.R
\name{kernel.pca.permute}
\alias{kernel.pca.permute}
\title{Assess variable importance}
\usage{
kernel.pca.permute(kpca.result, ncomp = 1, ..., directory = NULL)
}
\arguments{
\item{kpca.result}{a kernel.pca object returned by the
\code{\link{kernel.pca}} function.}

\item{ncomp}{integer. Number of KPCA components used to compute the 
importance. Default: \code{1}.}

\item{...}{list of character vectors. The parameter name must be the kernel 
name to be considered for permutation of variables. Provided vectors length 
has to be equal to the number of variables of the input dataset. A kernel is 
performed on each unique variables values. Crone-Crosby distances are 
computed on each KPCA performed on resulted kernels or meta-kernels and can 
be displayed using the \code{\link{plotVar.kernel.pca}}.}

\item{directory}{character. To limit computational burden, this argument 
allows to store / read temporary computed kernels.}
}
\value{
\code{kernel.pca.permute} returns a copy of the input 
\code{kpca.result} results and add values in the three entries: 
\code{cc.distances}, \code{cc.variables} and \code{cc.blocks}.
}
\description{
Assess importance of variables on a given PC component by computing the 
Crone-Crosby distance between original sample positions and sample positions 
obtained by a random permutation of the variables.
}
\details{
\code{plotVar.kernel.pca} produces a barplot for each block. The variables 
for which the importance has been computed with 
\code{\link{kernel.pca.permute}} are displayed. The representation is limited 
to the \code{ndisplay} most important variables.
}
\examples{
data(TARAoceans)

# compute one kernel for the psychem dataset
phychem.kernel <- compute.kernel(TARAoceans$phychem, kernel.func = "linear")
# perform a KPCA
kernel.pca.result <- kernel.pca(phychem.kernel)

# compute importance for all variables in this kernel
kernel.pca.result <- kernel.pca.permute(kernel.pca.result, 
                                        phychem = colnames(TARAoceans$phychem))

}
\references{
Mariette J. and Villa-Vialaneix N. (2018). Unsupervised multiple 
kernel learning for heterogeneous data integration. \emph{Bioinformatics}, 
\bold{34}(6), 1009-1015. DOI: \doi{10.1093/bioinformatics/btx682}

Crone L. and Crosby D. (1995). Statistical applications of a metric on 
subspaces to satellite meteorology. \emph{Technometrics}, \bold{37}(3), 
324-328.
}
\seealso{
\code{\link{compute.kernel}}, \code{\link{kernel.pca}}
}
\author{
Jerome Mariette <jerome.mariette@inrae.fr>
Nathalie Vialaneix <nathalie.vialaneix@inrae.fr>
}
